+++ src/main/java/io/kontur/eventapi/resource/EventResource.java
@@ -2,6 +2,7 @@ package io.kontur.eventapi.resource;
 
 import io.kontur.eventapi.entity.*;
 import io.kontur.eventapi.resource.dto.DataPaginationDTO;
+import io.kontur.eventapi.resource.dto.GeoJsonPaginationDTO;
 import io.kontur.eventapi.resource.dto.DateTimeRange;
 import io.kontur.eventapi.resource.dto.EpisodeFilterType;
 import io.kontur.eventapi.resource.dto.EventDto;
@@ -32,6 +33,7 @@ import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
@@ -114,6 +116,70 @@ public class EventResource {
             return ResponseEntity.noContent().build();
         }
         return ResponseEntity.ok(new DataPaginationDTO(events, events.get(events.size() - 1).getUpdatedAt()));
+    }
+
+    @GetMapping(path = "/geojson/events", produces = {APPLICATION_JSON_VALUE})
+    @Operation(tags = "Events", summary = "returns events in GeoJson format", description = "Returns events for specified feed name. All events are sorted by update date. <br> This method returns results using a cursor-based pagination approach:" +
+            "<ul><li>It accepts after and limit parameters.</li>" +
+            "<li>If you don't pass an after parameter the default value retrieves the first portion (or \"page\") of results.</li>" +
+            "<li>Paginated responses include a top-level responseMetadata object that includes a nextAfterValue.</li>" +
+            "<li>On your next call to the same method, set the after parameter equal to the nextAfterValue value you received on the last request to retrieve the next portion of the collection.</li>" +
+            "<li>nextAfterValue equals to the latest updatedAt event value on the page.</li></ul>")
+    @ApiResponse(responseCode = "200", description = "Successful operation", content = @Content(mediaType = "application/json", schema = @Schema(implementation = GeoJsonPaginationDTO.class)))
+    @ApiResponse(responseCode = "204", description = "No content. Try to check filters values.", content = @Content())
+    @PreAuthorize("hasAuthority('read:feed:'+#feed)")
+    public ResponseEntity<GeoJsonPaginationDTO> searchEventsGeoJson(
+            @Parameter(description = "Feed name")
+            @RequestParam(value = "feed")
+            String feed,
+            @Parameter(description = "Filters events by type. More than one can be chosen at once")
+            @RequestParam(value = "types", defaultValue = "")
+            List<EventType> eventTypes,
+            @Parameter(description = "Filters events by severity. More than one can be chosen at once")
+            @RequestParam(value = "severities", defaultValue = "")
+            List<Severity> severities,
+            @Parameter(description = "Includes events that were updated after this time. `updatedAt` property is used for selection. A date-time in ISO8601 format (e.g. \\\"2020-04-12T23:20:50.52Z\\\")")
+            @RequestParam(value = "after", required = false)
+            @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+            OffsetDateTime updatedAfter,
+            @Parameter(schema = @Schema(type = "string"),
+                    description = "Either a date-time or an interval, open or closed. Date and time expressions adhere to RFC 3339. Open intervals are expressed using double-dots. Examples:" +
+                            "<ul><li>A date-time: \"2018-02-12T23:20:50Z\"</li>" +
+                            "<li>A closed interval: \"2020-01-01T00:00:00Z/2020-12-01T00:00:00Z\"</li>" +
+                            "<li>Open intervals: \"2020-01-01T00:00:00Z/..\" or \"../2020-12-01T00:00:00Z\"</li></ul>" +
+                            "Only events that have a `startedAt` - `endedAt` interval that intersects the value of `datetime` are selected.")
+            @RequestParam(value = "datetime", required = false)
+            DateTimeRange datetime,
+            @Parameter(description = "Only hazards that have a geometry that intersects the bounding box are selected. The bounding box is provided as four numbers" +
+                    "<ul><li>Lower left corner, coordinate axis 1</li>" +
+                    "<li>Lower left corner, coordinate axis 2</li>" +
+                    "<li>Upper right corner, coordinate axis 1</li>" +
+                    "<li>Upper right corner, coordinate axis 2</li></ul>" +
+                    "The coordinate reference system of the values is WGS 84 longitude/latitude (http://www.opengis.net/def/crs/OGC/1.3/CRS84). For WGS 84 longitude/latitude the values are the sequence of minimum longitude, minimum latitude, maximum longitude and maximum latitude.")
+            @RequestParam(value = "bbox", required = false)
+            @ValidBbox
+            List<BigDecimal> bbox,
+            @Parameter(description = "Number of records on the page. Default value is 20, minimum - 1, maximum - 1000", example = "20", schema = @Schema(allowableValues = {}, minimum = "1", maximum = "1000"))
+            @RequestParam(value = "limit", defaultValue = "20")
+            @Min(1)
+            @Max(1000)
+            int limit,
+            @Parameter(description = "Sort selection. Default value is ASC")
+            @RequestParam(value = "sortOrder", defaultValue = "ASC")
+            SortOrder sortOrder,
+            @Parameter(description = "Type of filtering for event episodes (by severity): " +
+                    "<ul><li>ANY - one or more episodes match filters, all episodes are returned</li>" +
+                    "<li>LATEST - the latest episode matches filters, the latest episode is returned</li></ul>")
+            @RequestParam(value = "episodeFilterType", defaultValue = "ANY")
+            EpisodeFilterType episodeFilterType) {
+        Optional<GeoJsonPaginationDTO> paginationDTO = eventResourceService.searchEventsGeoJson(feed, eventTypes,
+                datetime != null && datetime.getFrom() != null ? datetime.getFrom() : null,
+                datetime != null && datetime.getTo() != null ? datetime.getTo() : null,
+                updatedAfter, limit, severities, sortOrder, bbox, episodeFilterType);
+        if (paginationDTO.isEmpty()) {
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.ok(paginationDTO.get());
     }
 
     @GetMapping(path = "/observations/{observationId}", produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})